+++ rbx_reflection/CHANGELOG.md
@@ -1,8 +1,10 @@
 # rbx_reflection Changelog
 
 ## Unreleased
+* Made `ReflectionDatabase` zero-copy. This involved removing `Cow`. ([#528])
 * Added support for one-to-many property migrations and added migrations for `UICorner.CornerRadius`. ([#612])
 
+[#528]: https://github.com/rojo-rbx/rbx-dom/pull/528
 [#612]: https://github.com/rojo-rbx/rbx-dom/pull/612
 
 ## 6.1.0 (2025-11-27)